tests: share the probe protocol fixture
A protocol the build genuinely has but that builds no device, so that
"score did not ask for a device" is a decision under test rather than an
absence. Used by the terminal-role tests on both sides of the addon
boundary.
